+++ src/de.sormuras.bach/main/java/de/sormuras/bach/project/SourceDirectory.java
@@ -17,14 +17,13 @@
 
 package de.sormuras.bach.project;
 
-import de.sormuras.bach.Scribe;
 import de.sormuras.bach.internal.Factory;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.regex.Pattern;
 
 /** A source directory potentially targeting a specific Java SE release. */
-public final class SourceDirectory implements Scribe {
+public final class SourceDirectory {
 
   private final Path path;
   private final int release;
@@ -65,11 +64,6 @@ public final class SourceDirectory implements Scribe {
   //
   // Normal API
   //
-
-  @Override
-  public void scribe(Scroll scroll) {
-    scroll.add("new SourceDirectory", path, release);
-  }
 
   public boolean isTargeted() {
     return release != 0;