+++ src/de.sormuras.bach/main/java/de/sormuras/bach/project/SourceUnit.java
@@ -17,7 +17,6 @@
 
 package de.sormuras.bach.project;
 
-import de.sormuras.bach.Scribe;
 import de.sormuras.bach.internal.Factory;
 import de.sormuras.bach.internal.Modules;
 import de.sormuras.bach.internal.Paths;
@@ -31,7 +30,7 @@ import java.util.List;
  *
  * @see ModuleDescriptor
  */
-public final class SourceUnit implements Comparable<SourceUnit>, Scribe {
+public final class SourceUnit implements Comparable<SourceUnit> {
 
   private final ModuleDescriptor descriptor;
   private final SourceDirectoryList sources;
@@ -87,15 +86,6 @@ public final class SourceUnit implements Comparable<SourceUnit>, Scribe {
   @Override
   public int compareTo(SourceUnit other) {
     return name().compareTo(other.name());
-  }
-
-  @Override
-  public void scribe(Scroll scroll) {
-    scroll.append("new SourceUnit(");
-    scroll.addNewLineAndContinue().add(descriptor).append(",");
-    scroll.addNewLineAndContinue().add(sources).append(",");
-    scroll.addNewLineAndContinue().add(resources);
-    scroll.append(")");
   }
 
   public String name() {