+++ src/de.sormuras.bach/main/java/de/sormuras/bach/project/TestSources.java
@@ -17,12 +17,11 @@
 
 package de.sormuras.bach.project;
 
-import de.sormuras.bach.Scribe;
 import de.sormuras.bach.internal.Factory;
 import de.sormuras.bach.internal.Factory.Kind;
 
 /** Source set of {@code test} modules. */
-public final class TestSources implements Realm<TestSources>, Scribe {
+public final class TestSources implements Realm<TestSources> {
 
   private final SourceUnitMap units;
 
@@ -51,11 +50,5 @@ public final class TestSources implements Realm<TestSources>, Scribe {
   @Factory(Kind.SETTER)
   public TestSources units(SourceUnitMap units) {
     return new TestSources(units);
-  }
-
-  @Override
-  public void scribe(Scroll scroll) {
-    scroll.append("TestSources.of()");
-    scroll.addNewLine().add(".units", units);
   }
 }